# Humoud-online-ordering-api-swagger-01.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
    "/users/v1/{userId}/profile": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Create a user profile",
        "description": "Create a profile for an existing user",
        "operationId": "createProfile",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Profile successfully created",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Invalid profile data"
          }
        }
      }
    },
    "/users/v1/register": {
      "post": {
        "tags": [
          "Users API"
        ],
        "summary": "Register a new user",
        "description": "Create a new user account",
        "operationId": "createUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "User successfully created",
            "content": {
              "application/json": {}
            }
          },
          "400": {
            "description": "Invalid user data"
          }
        }
      }
    },
...
    "/users/v1/{userId}": {
      "get": {
        "tags": [
          "Users API"
        ],
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID",
        "operationId": "user",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProfileRequest": {
        "required": [
          "firstName",
          "lastName"
        ],
        "type": "object",
        "properties": {
          "firstName": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "lastName": {
            "pattern": "^[a-zA-Z]+$",
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^\\d{8}$",
            "type": "string"
          }
        }
      },
      "UserRequest": {
        "required": [
          "age",
          "name",
          "password",
          "role",
          "username"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "USER",
              "ADMIN"
            ]
          }
        }
      },
...
      "User": {
        "required": [
          "id",
          "name",
          "username"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      }
    }
  }
}